requirements.txt
Django==3.2
djangorestframework==3.14
psycopg2>=2.8.6,<2.9
Pillow>=8.2.0,<8.3.0
django-filter==23.1
blurhash-python==1.2
